Adapting to EConstr.
This is not so easy, and it is not clear that when to use Term.constr
and when to use EConstr.constr was not at best.

Basically, we used Term.constr for indexing subterms in the hash
table, but at the price of fully normalizing the evars beforehand,
which may mean some extra cost (though it is done only once, so linear
in size if I'm correct).